fix(codemirror): resolve ref access error in main chat input [skip ci]
