+++ gemini-logs/fix-main-chat-input-ref-error.md
@@ -1,0 +1,23 @@
+## Error: Cannot access refs during render in `main-chat-input.tsx`
+
+### Error Description
+
+The `main-chat-input.tsx` component was throwing a `Cannot access refs during render` error. This was caused by creating CodeMirror extensions inline in the `extensions` prop of the `CodeMirror` component. The extensions were accessing refs and calling functions that depended on component state, which is not allowed during the render phase.
+
+### Fix Description
+
+To fix this error, I created a custom hook called `useCodeMirrorExtensions`. This hook encapsulates the logic for creating the CodeMirror extensions and ensures that they are only recreated when their dependencies change. The hook takes the `settings`, `addFiles`, `submitMessage`, and `isMobile` as arguments and returns a memoized array of extensions.
+
+By using this hook, I was able to remove the inline extension creation from the `MainChatInput` component, which resolved the `Cannot access refs during render` error.
+
+### Alignment with PHILOSOPHY.md
+
+This fix aligns with the `PHILOSOPHY.md` file in the following ways:
+
+- **Simplicity:** The custom hook simplifies the `MainChatInput` component by encapsulating the complex logic for creating the CodeMirror extensions.
+- **Clarity:** The code is now clearer and easier to understand, as the extension creation logic is separated from the component's render method.
+- **Efficiency:** The use of `useMemo` in the custom hook ensures that the extensions are only recreated when their dependencies change, which improves the component's performance.
+
+### Alignment with React Docs
+
+This fix aligns with the React documentation for `useRef`, which states that you should not read or write to `ref.current` during the render phase. By moving the extension creation to a custom hook, I am no longer accessing the refs during render.
